Remove override from clans
